+++ activity/jwt/examples/api/main.go
@@ -2,12 +2,11 @@ package main
 
 import (
 	"github.com/project-flogo/core/engine"
-	"github.com/project-flogo/microgateway/activity/jwt"
-	"github.com/project-flogo/microgateway/activity/jwt/example"
+	"github.com/project-flogo/microgateway/activity/jwt/examples"
 )
 
 func main() {
-	e, err := example.Example(&jwt.Activity{})
+	e, err := examples.Example()
 	if err != nil {
 		panic(err)
 	}
