+++ activity/sqld/examples/activity_example.go
@@ -1,21 +1,21 @@
-package example
+package examples
 
 import (
 	"github.com/project-flogo/contrib/activity/rest"
 	trigger "github.com/project-flogo/contrib/trigger/rest"
-	"github.com/project-flogo/core/activity"
 	"github.com/project-flogo/core/api"
 	"github.com/project-flogo/core/engine"
 	"github.com/project-flogo/microgateway"
+	"github.com/project-flogo/microgateway/activity/sqld"
 	microapi "github.com/project-flogo/microgateway/api"
 )
 
 // Example returns an API example
-func Example(activity activity.Activity) (engine.Engine, error) {
+func Example() (engine.Engine, error) {
 	app := api.NewApp()
 
 	gateway := microapi.New("Update")
-	serviceSQLD := gateway.NewService("SQLSecurity", activity)
+	serviceSQLD := gateway.NewService("SQLSecurity", &sqld.Activity{})
 	serviceSQLD.SetDescription("Look for sql injection attacks")
 
 	serviceUpdate := gateway.NewService("PetStorePetsUpdate", &rest.Activity{})
